+++ src/plugins/intel_cpu/src/nodes/experimental_detectron_generate_proposals_single_image.h
@@ -47,8 +47,8 @@ private:
     const int OUTPUT_SCORES{1};
 
     float min_size_;
-    int pre_nms_topn_;
-    int post_nms_topn_;
+    size_t pre_nms_topn_;
+    size_t post_nms_topn_;
     float nms_thresh_;
     float coordinates_offset;
